Move `locate_cmp_ops` to `invalid_literal_comparisons`
